erp_web/requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-session==0.8.0
psycopg2-binary==2.9.9
supabase==2.4.6
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==22.0.0
openpyxl==3.1.2
